#493 refactor
