+++ quolance-api/src/main/java/com/quolance/quolance_api/controllers/TextGenerationController.java
@@ -1,14 +1,17 @@
 package com.quolance.quolance_api.controllers;
 
-import com.quolance.quolance_api.dtos.text.GenerateTextDto;
 import com.quolance.quolance_api.dtos.text.GenerateApplicationLetterDto;
+import com.quolance.quolance_api.dtos.text.GenerateTextDto;
 import com.quolance.quolance_api.entities.User;
 import com.quolance.quolance_api.entities.enums.PromptType;
-import com.quolance.quolance_api.services.text.TextGenerationService;
+import com.quolance.quolance_api.services.ai_models.text.TextGenerationService;
 import com.quolance.quolance_api.util.SecurityUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/api/text-generation")
@@ -42,7 +45,7 @@ public class TextGenerationController {
     @PostMapping("/application")
     public ResponseEntity<String> generateApplicationText(@RequestBody GenerateApplicationLetterDto dto) {
         User user = SecurityUtil.getAuthenticatedUser();
-        // Include projectId for context in the final prompt
+
         String finalPrompt = "Project ID: " + dto.getProjectId() + ". " + dto.getPrompt();
 
         String generatedText = textGenerationService.generateText(
