+++ quolance-api/src/main/java/com/quolance/quolance_api/services/ai_models/text/prompts/ApplicationPromptStrategy.java
@@ -1,7 +1,7 @@
-package com.quolance.quolance_api.services.text.prompts;
+package com.quolance.quolance_api.services.ai_models.text.prompts;
 
-import com.quolance.quolance_api.entities.Project;
 import com.quolance.quolance_api.entities.Profile;
+import com.quolance.quolance_api.entities.Project;
 import com.quolance.quolance_api.entities.User;
 import com.quolance.quolance_api.entities.enums.PromptType;
 import com.quolance.quolance_api.entities.enums.Tag;
@@ -33,10 +33,8 @@ public class ApplicationPromptStrategy implements PromptStrategy {
 
     @Override
     public String generatePrompt(User user, String userPrompt) {
-        // 1) Parse the project ID from the user prompt and remove it from the text.
         ParsedPrompt parsed = parseProjectId(userPrompt);
 
-        // 2) Build project context (title, description, etc.), skipping the project ID.
         StringBuilder projectContext = new StringBuilder();
         if (parsed.projectId != null) {
             Optional<Project> projectOpt = projectRepository.findById(parsed.projectId);
@@ -65,12 +63,9 @@ public class ApplicationPromptStrategy implements PromptStrategy {
                     projectContext.append("Tags: ").append(project.getTags()).append(". ");
                 }
             }
-            // If not found, no project context is appended.
         }
 
-        // 3) Build freelancer (applicant) details.
         Profile profile = user.getProfile();
-        // Clearly label that these details are of the freelancer applying for the project.
         StringBuilder userDetails = new StringBuilder("Freelancer Details: ");
         appendIfNotBlank(userDetails, "First Name: ", user.getFirstName());
         appendIfNotBlank(userDetails, "Last Name: ", user.getLastName());
@@ -127,16 +122,11 @@ public class ApplicationPromptStrategy implements PromptStrategy {
             }
         }
 
-        // 4) Combine all elements: base prompt, project context, freelancer details, and the remaining text (with project ID removed)
         String finalPrompt = basePrompt + projectContext + userDetails + parsed.remainingPrompt;
         return finalPrompt;
     }
 
-    /**
-     * Uses a regex to extract a project ID from the user prompt.
-     * Supports both "ProjectID=" and "Project ID:" formats.
-     * The matched substring is removed from the final prompt.
-     */
+
     private ParsedPrompt parseProjectId(String userPrompt) {
         Pattern pattern = Pattern.compile("(ProjectID=|Project\\s*ID:)\\s*([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12})");
         Matcher matcher = pattern.matcher(userPrompt);
@@ -144,30 +134,23 @@ public class ApplicationPromptStrategy implements PromptStrategy {
             String idString = matcher.group(2);
             try {
                 UUID projectId = UUID.fromString(idString);
-                // Remove the matched substring from the prompt.
                 String remainingPrompt = matcher.replaceFirst("").trim();
                 return new ParsedPrompt(projectId, remainingPrompt);
             } catch (IllegalArgumentException e) {
-                // Not a valid UUID; return the original prompt.
                 return new ParsedPrompt(null, userPrompt);
             }
         }
-        // If no project ID is found, return the prompt unchanged.
         return new ParsedPrompt(null, userPrompt);
     }
 
-    /**
-     * Helper method to append a label and its value if the value is non-null and non-blank.
-     */
+
     private void appendIfNotBlank(StringBuilder sb, String label, String value) {
         if (value != null && !value.isBlank()) {
             sb.append(label).append(value).append(". ");
         }
     }
 
-    /**
-     * A simple class to hold the parsed project ID and the remaining user prompt.
-     */
+
     private static class ParsedPrompt {
         final UUID projectId;
         final String remainingPrompt;
