+++ quolance-api/src/main/java/com/quolance/quolance_api/services/ai_models/text/prompts/PromptStrategyRegistry.java
@@ -1,4 +1,4 @@
-package com.quolance.quolance_api.services.text.prompts;
+package com.quolance.quolance_api.services.ai_models.text.prompts;
 
 import com.quolance.quolance_api.entities.enums.PromptType;
 import org.springframework.stereotype.Component;
